In signals see if a file even exists before trying to delete.

+++ django_pdf_filler/signals.py
@@ -6,7 +6,8 @@ from .models import Document, Page
 
 def clean_up_document_on_delete(instance, **kwargs):
     try:
-        instance.file.delete(save=False)
+        if instance.file:
+            instance.file.delete(save=False)
     except:
         pass
 post_delete.connect(clean_up_document_on_delete, sender=Document)
@@ -14,7 +15,8 @@ post_delete.connect(clean_up_document_on_delete, sender=Document)
 
 def clean_up_layout_images_on_delete(instance, **kwargs):
     try:
-        instance.image.delete(save=False)
+        if instance.image:
+            instance.image.delete(save=False)
     except:
         pass
 post_delete.connect(clean_up_layout_images_on_delete, sender=Page)
